feat(membership): exact per-periodicity member plan pricing
Adds MemberPlan.periodicityPricing (nullable Json, additive migration) holding
exact min/target/max amounts in cents per payment period for every enabled
periodicity. Subscription monthlyAmount becomes a GraphQL Float so the derived
monthly value can carry fractions (e.g. 500.00/year -> 4166.67 cents/month);
all invoice-creating paths now round the period total through a single
calculatePeriodAmount helper, making any period price exactly representable.

- billing: rounding in legacy renew, periodic-job cron, Payrexx provider and
  crowdfunding revenue; importSubscription now multiplies by periodicity like
  every other path
- validation: public create/update checks the period amount against the
  per-periodicity minimum (fallback: derived from amountPerMonthMin)
- editor: per-periodicity min/target/max price panel on the member plan form
  (delta vs derived monthly price is displayed, not entered); subscription
  edit view edits the exact period amount and derives the fractional monthly
  value, so saving no longer truncates period-priced subscriptions
- website: subscribe defaults/validation/summary texts and member plan picker
  prefer the exact per-periodicity prices; upgrade + slider rounding
- csv export rounds monthlyAmount; regenerated schema-v2 + graphql clients

+++ libs/user-subscription/api/src/lib/subscription.model.ts
@@ -28,7 +28,7 @@ export class UpdateUserSubscriptionInput {
   @Field(() => PaymentPeriodicity)
   paymentPeriodicity: PaymentPeriodicityType;
 
-  @Field(() => Int)
+  @Field(() => Float)
   monthlyAmount: number;
 
   @Field()
@@ -64,7 +64,7 @@ export class CreateSubscriptionArgs {
   @Field(() => PaymentPeriodicity)
   paymentPeriodicity!: PaymentPeriodicity;
 
-  @Field(() => Int)
+  @Field(() => Float)
   monthlyAmount!: number;
 
   @Field({ nullable: true })
